added tsx files and removed old js files. still needs stronger type checking
