--- src/components/sidebar/SidebarContext.js
@@ -1,7 +1,0 @@
-'use client' 
-
-import React, {createContext} from 'react'
-
-const SidebarContext = createContext(null);
-
-export {SidebarContext}
